requirements.txt
aws-cdk-lib==2.201.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
aws-cdk.aws-ec2>=2.201.0
aws-cdk.aws-rds>=2.201.0
aws-cdk.aws-secretsmanager>=2.201.0
